Create sixth handler
Made a sixth (maj6 | dim7) handler. In the process of doing so, I had
also defined a new `Seventh` enum, and updated the `seventh` field in
the `Chord` struct. I also added a `Dim7` variant in the `Interval`
enum. Finally I wrote unit tests for the sixth handler, all passed.

This time I did the following:

- wrote a sixth handler for maj6 and dim7 intervals
- defined a new `Seventh` enum and updated `Chord`'s `seventh` field
- added a `Dim7` variant to the `Interval` enum